feat(Wbs): wbs 추가 기능 수정
