+++ project-template/src/main/java/org/teamone/projecttemplate/command/repository/CommandWbsRepository.java
@@ -4,6 +4,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import org.teamone.projecttemplate.command.entity.CommandWbs;
 
+import java.util.List;
+
 @Repository
 public interface CommandWbsRepository extends JpaRepository<CommandWbs, Integer> {
+    List<CommandWbs> findByProjectId(int projectId);
 }
